+++ index.d.ts
@@ -19,6 +19,7 @@ export interface StoreChannelConfig<K extends string, V extends StoreChannelObje
   schema(): V;
   size?: number;
   expiry?: number;
+  migrate?(link: V): void;
   destroy?(error: DOMException | DOMError, event: Event | null): boolean;
 }
 export interface StoreChannelObject<K extends string> {